Stop installing test resources
During colcon builds, colcon is using the CMAKE_INSTALL_PREFIX to
dictate where the files should be installed to. During Debian and RPM
builds, the DESTDIR environment variable is used. The latter approach
prefixes all of the install paths, including absolute paths like the
destination for the test test resource installation.

The end result is that they get installed into the staging directory for
the platform package, and typically end up in the package at some
unexpected location.

For RPM builds, this is a fatal build failure, as ROS packages are not
expected to install anything outside of /opt/ros/<foo>.

It doesn't appear that installing these resources is even necessary
because the working directory of the test means that they're already
available to the test that needs them without copying.

Signed-off-by: Scott K Logan <logans@cottsay.net>