Do not crash if there's an error querying endpoint info
I've observed in larger systems sometimes the RMW layer will return an error when trying to query publisher info.
If this happens, rclcpp throws an RCLError exception and the domain bridge crashes. I have not been able to
reliable reproduce this exception, but it seems to happen during the discovery phase. I've only observed it using rmw_connext_cpp.

In any case, it seems reasonable to log an error instead of throwing an exception so that the domain bridge threads can keep working.

Signed-off-by: Jacob Perron <jacob@openrobotics.org>