Define a `rmw_time_t` to `rclcpp::Duration` func
This commit introduces a utility header that defines a conversion
function from `rmw_time_t` to `rclcpp::Duration`.

This conversion function is necessary because `rmw_time_t` holds time
values (both sec/nsec) as `uint64_t` fields, but `rclcpp::Duration`
has a constructor that accepts sec as a `int32_t` value, raising the
concern of a narrowing conversion.

Similar in spirit to a function with a similar signature introduced in
https://github.com/ros2/rclcpp/pull/1467.

Signed-off-by: Abrar Rahman Protyasha <aprotyas@u.rochester.edu>